[VIM] For ignore tags files in ':grep' command
Problem: When the CWD is not a git repository, running ':grep' command
will also search for (and generally find) some matching in the ctags
generated file 'tags'. In case of git repository, the ingore rule is
given by the current '.gitignore' file.

Solution: Force exclusion of the 'tags' file using the RipGrep '--glob'
option.